requirements.txt
flask
flask_restplus
tensorflow
keras
numpy
Pillow
werkzeug
mysqlclient
